// LDC.Atlas.WebApp/Client/app/shared/dtos/section-search-result.ts
export interface SectionSearchResult {
    companyId: string;
    physicalContractCode: string;
    contractLabel: string;
    sectionId: number;
    contractDate: Date;
    traderDisplayName: string;
    status: string;
    departmentCode: string;
    departmentDescription: string;
    price: number;
    priceCode: string;
    priceUnitDescription: string;
    quantity: number;
    weightUnitCode: string;
    weightUnitDescription: string;
    contractQuantity: 0;
    currencyCode: string;
    currencyDescription: string;
    commodity1: string;
    commodity2: string;
    commodity3: string;
    commodity4: string;
    commodity5: string;
    contractTermCode: string;
    contractTermDescription: string;
    contractTermLocationPortCode: string;
    contractTermLocationDescription: string;
    paymentTermCode: string;
    paymentTermDescription: string;
    positionMonth: Date;
    sellerCode: string;
    sellerDescription: string;
    buyerCode: string;
    buyerDescription: string;
    counterparty: string;
    deliveryPeriodStart: Date;
    deliveryPeriodEnd: Date;
    charterReference: string;
    charterDescription: string;
    allocatedContractReference: string;
    contractType: string;
    createdBy: string;
    commodityId: number;
    originalQuantity: number;
    contractValue: number;
    memo: string;
    commodityDescription: string;
    cropYear: number;
    arbitrationCode: string;
    arbitrationDescription: string;
    periodType: string;
    positionType: string;
    portOfOrigin: string;
    portOfOriginDescription: string;
    portOfDestination: string;
    portOfDestinationDescription: string;
    groupingNumber: number;
    mainInvoiceReference: string;
    mainInvoiceDate: Date;
    percentageInvoiced: number;
    quantityCodeInvoiced: string;
    invoiceValue: number;
    paymentDate: Date;
    otherReference: string;
    quantityInvoiced: number;
    invoicingStatus: string;
    amendedBy: string;
    amendedOn: Date;
    vesselName: string;
    bLReference: string;
    charterManager: string;
    counterpartyRef: string;
    contractIssuedOn: Date;
    contractSentDate: Date;
    contractReturnedDate: Date;
    lastEmailReceivedDate: Date;
    shipmentPeriod: string;
    isTradeClosed: string;
    isTradeCancelled: string;
    contractInvoiceType: string;

}